Move all the new unit conversion bits into UnitHelper
