+++ include/forcing/DataProvider.hpp
@@ -3,8 +3,6 @@
 
 #include <string>
 #include <vector>
-#include <set>
-#include <tuple>
 #include <boost/core/span.hpp>
 
 namespace data_access
@@ -103,25 +101,6 @@ namespace data_access
         private:
     };
 
-    struct unit_error_log_key {
-        std::string requester_name;
-        std::string requester_variable;
-        std::string provider_name;
-        std::string provider_variable;
-        std::string failure_message;
-
-        bool operator<(unit_error_log_key const& rhs) const {
-            return std::tie(requester_name, requester_variable, provider_name, provider_variable, failure_message)
-                 < std::tie(rhs.requester_name, rhs.requester_variable, rhs.provider_name, rhs.provider_variable, rhs.failure_message);
-        }
-
-        bool operator==(unit_error_log_key const& rhs) const {
-            return std::tie(requester_name, requester_variable, provider_name, provider_variable, failure_message)
-                == std::tie(rhs.requester_name, rhs.requester_variable, rhs.provider_name, rhs.provider_variable, rhs.failure_message);
-        }
-    };
-
-    extern std::set<unit_error_log_key> unit_errors_reported;
 }
